astro/pyproject.toml
[project]
name = "astro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyswisseph>=2.10.3.2",
    "requests>=2.32.5",
]
